Enable context isolation
Electron-pdf was already using a single ipcApi object to expose all
functionality required in the web page context. However, the ipcApi
object itself was referenced directly, without going through the
contextBridge API.

Here we ensure that context isolation is enabled and share the ipcApi
object with the website ("main world") context. Note that having context
isolation enabled is now the default, so this should be the safer option
moving forward.